requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
anndata==0.10.3
annotated-types==0.6.0
ansicon==1.89.0
anyio==3.7.1
array-api-compat==1.4
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.34.2
botocore==1.34.2
cell-gears==0.1.2
certifi==2022.12.7
charset-normalizer==2.1.1
chex==0.1.7
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.2.0
croniter==1.4.1
cycler==0.12.1
datasets==2.15.0
dateutils==0.6.12
dcor==0.6
deepdiff==6.7.1
dill==0.3.7
dm-tree==0.1.8
docrep==0.3.2
etils==1.6.0
evaluate==0.4.1
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.9.0
flax==0.7.5
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.10.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.4
importlib-resources==6.1.1
inquirer==3.1.4
itsdangerous==2.1.2
jax==0.4.23
jaxlib==0.4.23
Jinja2==3.1.2
jinxed==1.2.1
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lightning==2.0.9.post0
lightning-cloud==0.5.57
lightning-utilities==0.10.0
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.1
mpmath==1.3.0
msgpack==1.0.7
mudata==0.2.3
multidict==6.0.4
multipledispatch==1.0.0
multiprocess==0.70.15
natsort==8.4.0
nest-asyncio==1.5.8
networkx==3.0
numba==0.58.1
numpy==1.24.1
numpyro==0.13.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.8
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
patsy==0.5.4
Pillow==9.3.0
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.17.2
PyJWT==2.8.0
pynndescent==0.5.11
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.8.6
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
regex==2023.10.3
requests==2.28.1
responses==0.18.0
rich==13.7.0
s3transfer==0.9.0
safetensors==0.4.1
scanpy==1.9.6
scikit-learn==1.3.2
scikit-misc==0.3.1
scipy==1.11.4
scvi-tools==1.0.4
seaborn==0.12.2
session-info==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sparse==0.14.0
starlette==0.27.0
starsessions==1.3.0
statsmodels==0.14.1
stdlib-list==0.10.0
sympy==1.12
tensorstore==0.1.51
threadpoolctl==3.2.0
tokenizers==0.15.0
toolz==0.12.0
torch==2.1.2+cu118
torch_geometric==2.4.0
torchaudio==2.1.2+cu118
torchmetrics==1.2.1
torchvision==0.16.2+cu118
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
umap-learn==0.5.5
urllib3==1.26.13
uvicorn==0.24.0.post1
wcwidth==0.2.12
websocket-client==1.7.0
websockets==12.0
xarray==2023.12.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
